% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfert_all_video.R
\name{tranfert_all_video}
\alias{tranfert_all_video}
\title{Tranfert all video}
\usage{
tranfert_all_video(
  videos = list.files(path = "~", pattern = "(mp4|mkv)", full.names = TRUE)
)
}
\arguments{
\item{videos}{liste de chemin de video}
}
\description{
Tranfert all video
}
