% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfert_all_video.R
\name{video_transfert}
\alias{video_transfert}
\title{Tranfert une video}
\usage{
video_transfert(une_video, path_video = "/home/cervan/video")
}
\arguments{
\item{une_video}{chemin vers une video}

\item{path_video}{chemin ou copier la video}
}
\description{
Tranfert une video
}
